# toeic-learning/api/swagger/swagger.yaml
swagger: "2.0"
info:
  version: "0.0.1"
  title: Toeic
# during dev, should point to your local machine
host: 103.114.107.16:8003
...
basePath: /api/learning
...
securityDefinitions:
  auth0:
    type: apiKey
    name: Authorization
    in: header
    description: "Authorization logged to application"
schemes:
  - http
  - https
# format of bodies a client can send (Content-Type)
consumes:
  - application/json
# format of the responses to the client (Accepts)
produces:
  - application/json
paths:
...
  /vocabularyByDay:
    x-swagger-router-controller: vocabulary
    get:
      description: Returns lesson info with parameters lessonId
      operationId: getVocabularyByDay
      tags:
        - vocabulary
      security:
        - auth0: []
      responses:
        "200":
          description: Success
          schema:
            # a pointer to a definition
            $ref: "#/definitions/MessageResp"
        # responses may fall through to errors
        default:
          description: Error
          schema:
            $ref: "#/definitions/ErrorResponse"
  /vocabularies/type/{vocabularyType}:
    x-swagger-router-controller: vocabulary
    get:
      description: Returns vocabular by type
      operationId: getVocabularyByType
      tags:
        - vocabulary
      security:
        - auth0: []
      parameters:
        - in: path
          name: vocabularyType
          type: string
          required: true
          description: vocabulary type
      responses:
        "200":
          description: Success
          schema:
            # a pointer to a definition
            $ref: "#/definitions/MessageResp"
        # responses may fall through to errors
        default:
          description: Error
          schema:
            $ref: "#/definitions/ErrorResponse"
  /vocabularies/type:
    x-swagger-router-controller: vocabulary
    get:
      description: Returns vocabular by type
      operationId: getAllType
      tags:
        - vocabulary
      security:
        - auth0: []
      responses:
        "200":
          description: Success
          schema:
            # a pointer to a definition
            $ref: "#/definitions/MessageResp"
        # responses may fall through to errors
        default:
          description: Error
          schema:
            $ref: "#/definitions/ErrorResponse"
    post:
      description: Update lesson info
      operationId: addType
      tags:
        - vocabulary
      security:
        - auth0: []
      parameters:
        - in: body
          name: body
          description: parameters with type, vocabulary, mean.
          schema:
            type: object
            required:
              - type
            properties:
              type:
                type: string

      responses:
        "200":
          description: Success
          schema:
            # a pointer to a definition
            $ref: "#/definitions/MessageResp"
        # responses may fall through to errors
        default:
          description: Error
          schema:
            $ref: "#/definitions/ErrorResponse"
  /vocabularies/{vocabularyId}:
    x-swagger-router-controller: vocabulary
    get:
      description: Returns vocabular by type
      operationId: getVocabularyById
      tags:
        - vocabulary
      security:
        - auth0: []
      parameters:
        - in: path
          name: vocabularyId
          type: string
          required: true
          description: vocabulary type
      responses:
        "200":
          description: Success
          schema:
            # a pointer to a definition
            $ref: "#/definitions/MessageResp"
        # responses may fall through to errors
        default:
          description: Error
          schema:
            $ref: "#/definitions/ErrorResponse"
    delete:
      description: Returns vocabular by type
      operationId: deleteVocabulary
      tags:
        - vocabulary
      security:
        - auth0: []
      parameters:
        - in: path
          name: vocabularyId
          type: string
          required: true
          description: vocabulary type
      responses:
        "200":
          description: Success
          schema:
            # a pointer to a definition
            $ref: "#/definitions/MessageResp"
        # responses may fall through to errors
        default:
          description: Error
          schema:
            $ref: "#/definitions/ErrorResponse"
  /vocabularies:
    x-swagger-router-controller: vocabulary
    get:
      description: Returns all vocabulary
      operationId: getAllVocabulary
      tags:
        - vocabulary
      security:
        - auth0: []
      parameters:
        - in: query
          name: limit
          type: integer
          description: limit value return
        - in: query
          name: page
          type: integer
          description: page get
      responses:
        "200":
          description: Success
          schema:
            # a pointer to a definition
            $ref: "#/definitions/MessageResp"
        # responses may fall through to errors
        default:
          description: Error
          schema:
            $ref: "#/definitions/ErrorResponse"
    post:
      description: Update lesson info
      operationId: addVocabulary
      tags:
        - vocabulary
      security:
        - auth0: []
      parameters:
        - in: body
          name: body
          description: parameters with type, vocabulary, mean.
          schema:
            $ref: '#/definitions/VocabularyReq'
      responses:
        "200":
          description: Success
          schema:
            # a pointer to a definition
            $ref: "#/definitions/MessageResp"
        # responses may fall through to errors
        default:
          description: Error
          schema:
            $ref: "#/definitions/ErrorResponse"
...
definitions:
...
  VocabularyReq:
    required:
      - type
      - vocabulary
    properties:
      type:
        type: string
      vocabulary:
        type: string
      mean:
        type: string
...
  MessageResp:
    type: object
    required: 
      - success
    properties:
      success:
        type: boolean
      message:
        type: string
      value:
        type: object
  ErrorResponse:
    required:
      - message
    properties:
      message:
        type: string